Change deletePoll to delete request

+++ src/pages/api/deletePoll/[pollId].ts
@@ -18,9 +18,24 @@ export default async function deletePoll(
   req: NextApiRequest,
   res: NextApiResponse<Data>,
 ): Promise<void> {
-  const { pollId } = req.body;
-  // TODO: Add session check to verify it is coordinator. Also additional security step of verifying coordinator's signature before deleting poll
   try {
+    if (req.method !== 'DELETE') {
+      res.setHeader('Allow', 'DELETE');
+      return res
+        .status(405)
+        .json({ success: false, message: 'Method not allowed' });
+    }
+
+    const pollId = req.query.pollId;
+
+    if (typeof pollId !== 'string') {
+      return res.status(400).json({
+        success: false,
+        message: 'Invalid pollId',
+      });
+    }
+
+    // TODO: Add session check to verify it is coordinator. Also additional security step of verifying coordinator's signature before deleting poll
     await prisma.poll.delete({
       where: {
         id: BigInt(pollId),
